+++ popkorn/src/commonMain/kotlin/cc/popkorn/Platform.kt
@@ -1,5 +1,6 @@
 package cc.popkorn
 
+import cc.popkorn.core.Injector
 import cc.popkorn.pools.ResolverPool
 import kotlin.reflect.KClass
 
@@ -19,3 +20,5 @@ expect class WeakReference<T : Any>(referred: T) {
 internal expect fun <T : Any> KClass<T>.getName(): String
 
 internal expect fun <T : Any> KClass<T>.needsResolver(resolverPool: ResolverPool): Boolean
+
+internal expect fun createDefaultInjector(): Injector
